Updated Move Class
Changed move class to be called in the controller. Changed move class functionality into a class. Added an event that is called when drag is stopped that sets the class's coordinates within the controller.

Removed unneeded code I wrote a while ago in GUIMainWindow.

Fixed a bug in Class Collection